ollamawebui-customizable/back-up-build versions/ollamawebui-customizable/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
requests
waitress
python-dotenv